feat: split item-list into two components
