+++ src/components/item-details/index.ts
@@ -1,0 +1,3 @@
+import ItemDetails from './item-details';
+
+export default ItemDetails;
