+++ src/components/item-list/item-list.css
@@ -25,11 +25,6 @@
   justify-content: space-evenly;
 }
 
-.section-right {
-  flex-basis: 50%;
-  flex-shrink: 0;
-}
-
 .dwarf {
   background-image: url('../../assets/icons/dwarf.png');
 }
